feat(freshagent): tab routes on the LayoutStore — order, selection, next/prev, title-aware has (AUTO-03)
