+++ crates/freshell-freshagent/src/pane_ops.rs
@@ -42,10 +42,9 @@ use uuid::Uuid;
 
 use freshell_protocol::{ServerMessage, UiCommand};
 
+use crate::layout_store::RenameOutcome;
 use crate::terminal_tabs::{spawn_terminal_pane, TerminalSpawnResult};
-use crate::{
-    approx_json, authorized, fail_json, ok_json, parse_required_name, FreshAgentState, TabRecord,
-};
+use crate::{approx_json, authorized, fail_json, ok_json, parse_required_name, FreshAgentState};
 
 /// Mount the pane + tab lifecycle routes onto an existing router. Split out of
 /// [`crate::router`] so `lib.rs`'s route table stays a single glance-able list;
@@ -323,7 +322,22 @@ fn drop_and_lookup_pane_tab(state: &FreshAgentState, pane_id: &str) -> Option<St
 
 // ── POST /api/tabs/:id/select ───────────────────────────────────────────────
 
-/// `POST /api/tabs/:id/select` (`router.ts:834-838`). Always broadcasts
+/// Fold a store tab mutation into the legacy `ok(result, result.message ||
+/// success)` envelope every tab route shares (`router.ts:836,848,854`): the
+/// data is the outcome itself — `{tabId}` on success, `{message}` otherwise —
+/// and the envelope message mirrors `result.message || <success>`.
+fn tab_outcome_response(outcome: RenameOutcome, success_message: &str) -> Response {
+    match outcome.tab_id {
+        Some(tab_id) => ok_json(json!({ "tabId": tab_id }), success_message),
+        None => {
+            let message = outcome.message.unwrap_or("tab not found");
+            ok_json(json!({ "message": message }), message)
+        }
+    }
+}
+
+/// `POST /api/tabs/:id/select` (`router.ts:834-838`): store `selectTab`
+/// (persists `activeTabId` — Task 14, AUTO-03). Always broadcasts
 /// `ui.command{tab.select}` regardless of whether the tab exists, matching the
 /// original exactly (`selectTab` returns `{message:'tab not found'}` for an
 /// unknown id, but the broadcast fires unconditionally either way).
@@ -336,25 +350,27 @@ pub(crate) async fn select_tab(
         return fail_json(StatusCode::UNAUTHORIZED, "unauthorized".to_string());
     }
 
-    let exists = state.tabs.lock().expect("tabs mutex").contains_key(&tab_id);
+    let outcome = state.layout.select_tab(&tab_id);
 
     state.broadcast(&ServerMessage::UiCommand(UiCommand {
         command: "tab.select".to_string(),
         payload: Some(json!({ "id": tab_id })),
     }));
 
-    if exists {
-        ok_json(json!({ "tabId": tab_id }), "tab selected")
-    } else {
-        ok_json(json!({ "message": "tab not found" }), "tab not found")
-    }
+    tab_outcome_response(outcome, "tab selected")
 }
 
 // ── PATCH /api/tabs/:id ─────────────────────────────────────────────────────
 
-/// `PATCH /api/tabs/:id` (`router.ts:840-849`): rename a tab. Legacy applies no
-/// length bound here (unlike `PATCH /api/panes/:id`'s `MAX_TERMINAL_TITLE_OVERRIDE_LENGTH`
-/// check) -- mirrored exactly, no bound added.
+/// `PATCH /api/tabs/:id` (`router.ts:840-849`): rename a tab on the store
+/// (`renameTab` — single-pane pane-title mirror included; `{message:'no layout
+/// snapshot'}` when the store is empty, Node parity for the no-client hole).
+/// Broadcasts `ui.command{tab.rename}` ONLY when the rename resolved
+/// (`router.ts:845`'s `if (result?.tabId)` guard). Legacy applies no length
+/// bound here (unlike `PATCH /api/panes/:id`'s `MAX_TERMINAL_TITLE_OVERRIDE_LENGTH`
+/// check) -- mirrored exactly, no bound added. The legacy `TabRecord.title`
+/// shadow is kept updated too: split/close/respawn continuity and restore
+/// still read it (see this module's top doc comment).
 pub(crate) async fn rename_tab(
     State(state): State<FreshAgentState>,
     Path(tab_id): Path<String>,
@@ -369,31 +385,35 @@ pub(crate) async fn rename_tab(
         return fail_json(StatusCode::BAD_REQUEST, "name required".to_string());
     };
 
-    let mut tabs = state.tabs.lock().expect("tabs mutex");
-    let Some(record) = tabs.get_mut(&tab_id) else {
-        drop(tabs);
-        return ok_json(json!({ "message": "tab not found" }), "tab not found");
-    };
-    record.title = Some(name.clone());
-    drop(tabs);
+    let outcome = state.layout.rename_tab(&tab_id, &name);
 
-    state.broadcast(&ServerMessage::UiCommand(UiCommand {
-        command: "tab.rename".to_string(),
-        payload: Some(json!({ "id": tab_id, "title": name })),
-    }));
+    if let Some(record) = state.tabs.lock().expect("tabs mutex").get_mut(&tab_id) {
+        record.title = Some(name.clone());
+    }
 
-    ok_json(json!({ "tabId": tab_id }), "tab renamed")
+    if outcome.tab_id.is_some() {
+        state.broadcast(&ServerMessage::UiCommand(UiCommand {
+            command: "tab.rename".to_string(),
+            payload: Some(json!({ "id": tab_id, "title": name })),
+        }));
+    }
+
+    tab_outcome_response(outcome, "tab renamed")
 }
 
 // ── DELETE /api/tabs/:id ─────────────────────────────────────────────────────
 
-/// `DELETE /api/tabs/:id` (`router.ts:851-855`): close a tab and every pane it
-/// owns. Always broadcasts `ui.command{tab.close}` regardless of whether the
-/// tab existed (matching `router.ts:853`'s unconditional broadcast, same
-/// pattern as `select_tab`). See this module's top doc comment: this removes
-/// ONLY local bookkeeping for every owned pane -- no `registry.kill` call, so
-/// each pane's terminal (if any) keeps running as a background session in the
-/// shared registry, exactly like the legacy `closeTab` does.
+/// `DELETE /api/tabs/:id` (`router.ts:851-855`): store `closeTab` (drives the
+/// response — `{tabId}` / `{message:'tab not found'|'no layout snapshot'}`),
+/// plus the pre-existing owned-resource cleanup on the legacy shadow maps
+/// (`tabs`/`pane_tabs`/`terminal_panes`/`content_panes` — split/close/respawn
+/// continuity still reads them). Always broadcasts `ui.command{tab.close}`
+/// regardless of whether the tab existed (matching `router.ts:853`'s
+/// unconditional broadcast, same pattern as `select_tab`). See this module's
+/// top doc comment: this removes ONLY local bookkeeping for every owned pane
+/// -- no `registry.kill` call, so each pane's terminal (if any) keeps running
+/// as a background session in the shared registry, exactly like the legacy
+/// `closeTab` does.
 pub(crate) async fn delete_tab(
     State(state): State<FreshAgentState>,
     Path(tab_id): Path<String>,
@@ -403,9 +423,18 @@ pub(crate) async fn delete_tab(
         return fail_json(StatusCode::UNAUTHORIZED, "unauthorized".to_string());
     }
 
-    let removed: Option<TabRecord> = state.tabs.lock().expect("tabs mutex").remove(&tab_id);
+    let outcome = state.layout.close_tab(&tab_id);
 
-    let (message, data) = if removed.is_some() {
+    // Legacy shadow cleanup, keyed on the legacy maps themselves (a tab can
+    // exist there without a store row and vice versa during the AUTO-03
+    // transition; the response above stays store-authoritative either way).
+    if state
+        .tabs
+        .lock()
+        .expect("tabs mutex")
+        .remove(&tab_id)
+        .is_some()
+    {
         let owned_panes: Vec<String> = state
             .pane_tabs
             .lock()
@@ -431,28 +460,24 @@ pub(crate) async fn delete_tab(
                 .expect("pane_tabs mutex")
                 .remove(&pane_id);
         }
-        ("tab closed", json!({ "tabId": tab_id }))
-    } else {
-        ("tab not found", json!({ "message": "tab not found" }))
-    };
+    }
 
     state.broadcast(&ServerMessage::UiCommand(UiCommand {
         command: "tab.close".to_string(),
         payload: Some(json!({ "id": tab_id })),
     }));
 
-    ok_json(data, message)
+    tab_outcome_response(outcome, "tab closed")
 }
 
 // ── GET /api/tabs/has ───────────────────────────────────────────────────
 
-/// `GET /api/tabs/has?target=` (`router.ts:857-861`): `{ exists }`. Legacy's
-/// `layoutStore.hasTab` matches by id OR title (ambiguous-title resolution);
-/// this port has no title-based lookup anywhere (the established precedent
-/// across every route in this module -- `select_pane`/`rename_tab`/etc. all
-/// resolve strictly by id via [`FreshAgentState::tabs`]/[`FreshAgentState::pane_tabs`]),
-/// so `target` is matched against tab id ONLY. A missing/empty `target`
-/// mirrors the original's `target ? ... : false` short-circuit.
+/// `GET /api/tabs/has?target=` (`router.ts:857-861`): `{ exists }` via the
+/// store's `hasTab`, which matches by tab id OR title (`layout-store.ts:336-339`
+/// — Task 14, AUTO-03 retires the id-only interim lookup). A missing/empty
+/// `target` mirrors the original's `target ? ... : false` short-circuit —
+/// normalized HERE, before the store call, so the store never sees a blank
+/// target (caller hygiene for `Some("")`).
 pub(crate) async fn tabs_has(
     State(state): State<FreshAgentState>,
     headers: HeaderMap,
@@ -462,30 +487,34 @@ pub(crate) async fn tabs_has(
         return fail_json(StatusCode::UNAUTHORIZED, "unauthorized".to_string());
     }
     let target = params.get("target").map(String::as_str).unwrap_or("");
-    let exists = !target.is_empty() && state.tabs.lock().expect("tabs mutex").contains_key(target);
+    let exists = !target.is_empty() && state.layout.has_tab(target);
     ok_json(json!({ "exists": exists }), "")
 }
 
-// ── POST /api/tabs/next, POST /api/tabs/prev (honest deferral) ─────────
+// ── POST /api/tabs/next, POST /api/tabs/prev ───────────────────────────
 
-/// `POST /api/tabs/next` / `POST /api/tabs/prev` (`router.ts:863-877`): cycle
-/// the active tab through an ORDERED tab list. Deferred: this port's
-/// [`FreshAgentState::tabs`] is an unordered `HashMap` with no active-tab-id
-/// concept at all -- `terminal_tabs::list_tabs` already hard-codes
-/// `"activeTabId": Value::Null` for `GET /api/tabs` (an established Slice 1
-/// reduced-fidelity precedent this route would need to break). Implementing
-/// real cycling needs an ordered tab sequence + an active-tab pointer added
-/// to the shared `FreshAgentState` struct in `lib.rs` -- out of this slice's
-/// owned-file scope (`pane_ops.rs` + route registration + `terminal_tabs.rs`
-/// only) while five other agents work concurrently in this same crate.
-/// Returns an honest 400 naming exactly this gap rather than silently
-/// no-op-ing or fabricating an ordering.
-const TAB_CYCLE_DEFERRAL_MESSAGE: &str = "tab cycling (next/prev) is not implemented on this \
-     server: it requires an ordered tab sequence + an active-tab id that FreshAgentState does \
-     not model (GET /api/tabs already reports activeTabId: null, an established Slice 1 \
-     precedent). Adding that state means extending FreshAgentState in lib.rs, which is out of \
-     this slice's owned-file scope. Deferred pending tab-ordering/active-tab state landing.";
+/// The shared next/prev response fold (`router.ts:863-877`): broadcast
+/// `ui.command{tab.select,{id}}` ONLY when a tab resolved (`if (result?.tabId)`),
+/// then `ok(result, result?.message || 'tab selected')` — `{tabId}` on a
+/// resolve, `{message:'no tabs'}` when the store has no snapshot or no tabs
+/// (`selectNextTab`/`selectPrevTab`, `layout-store.ts:589-607`).
+fn tab_cycle_response(state: &FreshAgentState, resolved: Option<String>) -> Response {
+    match resolved {
+        Some(tab_id) => {
+            state.broadcast(&ServerMessage::UiCommand(UiCommand {
+                command: "tab.select".to_string(),
+                payload: Some(json!({ "id": tab_id })),
+            }));
+            ok_json(json!({ "tabId": tab_id }), "tab selected")
+        }
+        None => ok_json(json!({ "message": "no tabs" }), "no tabs"),
+    }
+}
 
+/// `POST /api/tabs/next` (`router.ts:863-869`): ordered active-tab cycling on
+/// the shared LayoutStore (Task 14, AUTO-03 — the Slice 3b-1 honest-400
+/// deferral dies here: the store IS the ordered tab sequence + active-tab id
+/// that deferral was waiting on).
 pub(crate) async fn tabs_next(
     State(state): State<FreshAgentState>,
     headers: HeaderMap,
@@ -493,12 +522,11 @@ pub(crate) async fn tabs_next(
     if !authorized(&headers, &state.auth_token) {
         return fail_json(StatusCode::UNAUTHORIZED, "unauthorized".to_string());
     }
-    fail_json(
-        StatusCode::BAD_REQUEST,
-        TAB_CYCLE_DEFERRAL_MESSAGE.to_string(),
-    )
+    let resolved = state.layout.select_next_tab();
+    tab_cycle_response(&state, resolved)
 }
 
+/// `POST /api/tabs/prev` (`router.ts:871-877`): see [`tabs_next`].
 pub(crate) async fn tabs_prev(
     State(state): State<FreshAgentState>,
     headers: HeaderMap,
@@ -506,10 +534,8 @@ pub(crate) async fn tabs_prev(
     if !authorized(&headers, &state.auth_token) {
         return fail_json(StatusCode::UNAUTHORIZED, "unauthorized".to_string());
     }
-    fail_json(
-        StatusCode::BAD_REQUEST,
-        TAB_CYCLE_DEFERRAL_MESSAGE.to_string(),
-    )
+    let resolved = state.layout.select_prev_tab();
+    tab_cycle_response(&state, resolved)
 }
 
 // ── GET /api/layout/snapshot ────────────────────────────────────────────
@@ -933,1033 +959,9 @@ pub(crate) async fn swap_pane(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::body::Body;
-    use axum::http::Request;
-    use std::sync::Arc;
-    use tower::util::ServiceExt;
-
-    fn state_with_registry() -> FreshAgentState {
-        let (tx, _rx) = tokio::sync::broadcast::channel::<String>(64);
-        FreshAgentState::new(Arc::new("tok".to_string()), Arc::new(tx))
-            .with_terminal_registry(freshell_terminal::TerminalRegistry::new())
-    }
-
-    fn app(state: FreshAgentState) -> Router {
-        crate::router(state)
-    }
-
-    async fn body_json(resp: Response) -> Value {
-        let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-            .await
-            .unwrap();
-        serde_json::from_slice(&bytes).unwrap()
-    }
-
-    async fn post(router: Router, uri: &str, body: Value, auth: bool) -> (StatusCode, Value) {
-        let mut req = Request::builder()
-            .method("POST")
-            .uri(uri)
-            .header("content-type", "application/json");
-        if auth {
-            req = req.header("x-auth-token", "tok");
-        }
-        let resp = router
-            .oneshot(req.body(Body::from(body.to_string())).unwrap())
-            .await
-            .unwrap();
-        let status = resp.status();
-        (status, body_json(resp).await)
-    }
-
-    async fn patch(router: Router, uri: &str, body: Value, auth: bool) -> (StatusCode, Value) {
-        let mut req = Request::builder()
-            .method("PATCH")
-            .uri(uri)
-            .header("content-type", "application/json");
-        if auth {
-            req = req.header("x-auth-token", "tok");
-        }
-        let resp = router
-            .oneshot(req.body(Body::from(body.to_string())).unwrap())
-            .await
-            .unwrap();
-        let status = resp.status();
-        (status, body_json(resp).await)
-    }
-
-    async fn delete(router: Router, uri: &str, auth: bool) -> (StatusCode, Value) {
-        let mut req = Request::builder().method("DELETE").uri(uri);
-        if auth {
-            req = req.header("x-auth-token", "tok");
-        }
-        let resp = router
-            .oneshot(req.body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-        let status = resp.status();
-        (status, body_json(resp).await)
-    }
-
-    /// Create a real shell tab via the existing Slice-1 create route, returning
-    /// (tabId, paneId, terminalId).
-    async fn create_shell_tab(router: Router) -> (String, String, String) {
-        let tmp = std::env::temp_dir();
-        let (status, body) = post(
-            router,
-            "/api/tabs",
-            json!({ "mode": "shell", "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        (
-            body["data"]["tabId"].as_str().unwrap().to_string(),
-            body["data"]["paneId"].as_str().unwrap().to_string(),
-            body["data"]["terminalId"].as_str().unwrap().to_string(),
-        )
-    }
-
-    // ── auth ────────────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn split_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/split", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn close_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/close", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn select_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/select", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn select_tab_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/tabs/nope/select", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn rename_tab_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = patch(app(state), "/api/tabs/nope", json!({"name":"x"}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn delete_tab_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = delete(app(state), "/api/tabs/nope", false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    // ── split ───────────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn split_unknown_pane_is_404() {
-        let state = state_with_registry();
-        let (status, body) = post(
-            app(state),
-            "/api/panes/does-not-exist/split",
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::NOT_FOUND, "{body}");
-        assert_eq!(body["message"], json!("pane not found"));
-    }
-
-    #[tokio::test]
-    async fn split_agent_pane_is_honest_400() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_id, pane_id, _terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/split"),
-            json!({ "agent": "opencode" }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        let msg = body["message"].as_str().unwrap();
-        assert!(msg.contains("fresh-agent"), "{msg}");
-    }
-
-    #[tokio::test]
-    async fn split_terminal_pane_spawns_real_pty_and_broadcasts_pane_split() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, pane_id, _terminal_id) = create_shell_tab(router.clone()).await;
-        // Drain the tab.create broadcast so we only see this split's frame.
-        let _ = rx.recv().await;
-
-        let tmp = std::env::temp_dir();
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/split"),
-            json!({ "direction": "vertical", "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let new_pane_id = body["data"]["paneId"].as_str().unwrap().to_string();
-        let new_terminal_id = body["data"]["terminalId"].as_str().unwrap().to_string();
-        assert_ne!(new_pane_id, pane_id);
-        assert!(state
-            .terminal_registry
-            .clone()
-            .unwrap()
-            .is_running(&new_terminal_id));
-
-        let frame = rx.recv().await.expect("pane.split broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("pane.split"));
-        assert_eq!(msg["payload"]["tabId"], json!(tab_id));
-        assert_eq!(msg["payload"]["paneId"], json!(pane_id));
-        assert_eq!(msg["payload"]["direction"], json!("vertical"));
-        assert_eq!(msg["payload"]["newPaneId"], json!(new_pane_id));
-        assert_eq!(
-            msg["payload"]["newContent"]["terminalId"],
-            json!(new_terminal_id)
-        );
-
-        state
-            .terminal_registry
-            .clone()
-            .unwrap()
-            .kill(&new_terminal_id);
-    }
-
-    #[tokio::test]
-    async fn split_browser_pane_registers_cheap_content_no_terminal() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_id, pane_id, _terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/split"),
-            json!({ "browser": "https://example.com" }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert!(body["data"]["terminalId"].is_null());
-        assert_eq!(body["message"], json!("pane split (non-terminal)"));
-    }
-
-    // ── close ───────────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn close_unknown_pane_is_ok_with_not_found_message() {
-        let state = state_with_registry();
-        let (status, body) = post(
-            app(state),
-            "/api/panes/does-not-exist/close",
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("pane not found"));
-    }
-
-    #[tokio::test]
-    async fn close_only_pane_in_tab_is_refused() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/close"),
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("cannot close only pane"));
-        // Untouched: the pane is still resolvable and its terminal still runs.
-        assert!(state.pane_tabs.lock().unwrap().contains_key(&pane_id));
-        assert!(state
-            .terminal_registry
-            .clone()
-            .unwrap()
-            .is_running(&terminal_id));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    /// The required split-then-close lifecycle: split creates a second real
-    /// pane/PTY, close removes ONLY this crate's bookkeeping for it -- the PTY
-    /// keeps running in the shared registry (this module's documented
-    /// PTY-cleanup-parity finding: legacy never kills on pane close), so
-    /// there is no orphan (it remains tracked by the SAME registry every
-    /// other surface uses) and no leak of crate-local bookkeeping either.
-    #[tokio::test]
-    async fn split_then_close_removes_bookkeeping_but_keeps_pty_alive_no_orphan() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, first_pane_id, first_terminal_id) = create_shell_tab(router.clone()).await;
-
-        let tmp = std::env::temp_dir();
-        let (status, split_body) = post(
-            router.clone(),
-            &format!("/api/panes/{first_pane_id}/split"),
-            json!({ "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{split_body}");
-        let new_pane_id = split_body["data"]["paneId"].as_str().unwrap().to_string();
-        let new_terminal_id = split_body["data"]["terminalId"]
-            .as_str()
-            .unwrap()
-            .to_string();
-
-        let (status, close_body) = post(
-            router,
-            &format!("/api/panes/{new_pane_id}/close"),
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{close_body}");
-        assert_eq!(close_body["data"]["tabId"], json!(tab_id));
-
-        // Bookkeeping removed: the closed pane no longer resolves.
-        assert!(!state.pane_tabs.lock().unwrap().contains_key(&new_pane_id));
-        assert!(!state
-            .terminal_panes
-            .lock()
-            .unwrap()
-            .contains_key(&new_pane_id));
-
-        // No orphan PTY: registry state proves BOTH terminals are still
-        // tracked and running (background-session semantics, not a leak).
-        let registry = state.terminal_registry.clone().unwrap();
-        assert!(registry.is_running(&first_terminal_id));
-        assert!(registry.is_running(&new_terminal_id));
-
-        registry.kill(&first_terminal_id);
-        registry.kill(&new_terminal_id);
-    }
-
-    // ── select ──────────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn select_unknown_pane_is_ok_with_not_found_message_and_no_broadcast() {
-        let state = state_with_registry();
-        let mut rx = state.broadcast_tx.subscribe();
-        let (status, body) = post(
-            app(state),
-            "/api/panes/does-not-exist/select",
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("pane not found"));
-        assert!(
-            rx.try_recv().is_err(),
-            "must not broadcast for unresolved pane"
-        );
-    }
-
-    #[tokio::test]
-    async fn select_pane_resolves_tab_via_pane_tabs_and_broadcasts() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-        let _ = rx.recv().await; // drain tab.create
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/select"),
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["tabId"], json!(tab_id));
-        assert_eq!(body["data"]["paneId"], json!(pane_id));
-
-        let frame = rx.recv().await.expect("pane.select broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("pane.select"));
-        assert_eq!(msg["payload"]["tabId"], json!(tab_id));
-        assert_eq!(msg["payload"]["paneId"], json!(pane_id));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    // ── tab select ──────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn select_unknown_tab_still_broadcasts_but_reports_not_found() {
-        let state = state_with_registry();
-        let mut rx = state.broadcast_tx.subscribe();
-        let (status, body) = post(
-            app(state),
-            "/api/tabs/does-not-exist/select",
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("tab not found"));
-        let frame = rx.recv().await.expect("legacy-exact: always broadcasts");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("tab.select"));
-    }
-
-    #[tokio::test]
-    async fn select_known_tab_succeeds() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, _pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/tabs/{tab_id}/select"),
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["tabId"], json!(tab_id));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    // ── tab rename ──────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn rename_tab_missing_name_is_400() {
-        let state = state_with_registry();
-        let (status, body) = patch(app(state), "/api/tabs/does-not-exist", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        assert_eq!(body["message"], json!("name required"));
-    }
-
-    #[tokio::test]
-    async fn rename_unknown_tab_reports_not_found() {
-        let state = state_with_registry();
-        let (status, body) = patch(
-            app(state),
-            "/api/tabs/does-not-exist",
-            json!({"name":"New Name"}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("tab not found"));
-    }
-
-    #[tokio::test]
-    async fn rename_known_tab_broadcasts_tab_rename() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, _pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-        let _ = rx.recv().await; // drain tab.create
-
-        let (status, body) = patch(
-            router,
-            &format!("/api/tabs/{tab_id}"),
-            json!({"name":"Renamed"}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["tabId"], json!(tab_id));
-
-        let frame = rx.recv().await.expect("tab.rename broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("tab.rename"));
-        assert_eq!(msg["payload"]["id"], json!(tab_id));
-        assert_eq!(msg["payload"]["title"], json!("Renamed"));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    // ── tab delete ──────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn delete_unknown_tab_reports_not_found_but_still_broadcasts() {
-        let state = state_with_registry();
-        let mut rx = state.broadcast_tx.subscribe();
-        let (status, body) = delete(app(state), "/api/tabs/does-not-exist", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("tab not found"));
-        let frame = rx.recv().await.expect("legacy-exact: always broadcasts");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("tab.close"));
-    }
-
-    #[tokio::test]
-    async fn delete_tab_removes_tab_and_every_owned_pane_without_killing_ptys() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, first_pane_id, first_terminal_id) = create_shell_tab(router.clone()).await;
-
-        let tmp = std::env::temp_dir();
-        let (status, split_body) = post(
-            router.clone(),
-            &format!("/api/panes/{first_pane_id}/split"),
-            json!({ "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{split_body}");
-        let second_pane_id = split_body["data"]["paneId"].as_str().unwrap().to_string();
-        let second_terminal_id = split_body["data"]["terminalId"]
-            .as_str()
-            .unwrap()
-            .to_string();
-
-        let (status, body) = delete(router, &format!("/api/tabs/{tab_id}"), true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["tabId"], json!(tab_id));
-
-        assert!(!state.tabs.lock().unwrap().contains_key(&tab_id));
-        assert!(!state.pane_tabs.lock().unwrap().contains_key(&first_pane_id));
-        assert!(!state
-            .pane_tabs
-            .lock()
-            .unwrap()
-            .contains_key(&second_pane_id));
-
-        // No PTY kill on tab close (this module's documented parity finding) --
-        // both terminals remain tracked + running in the shared registry.
-        let registry = state.terminal_registry.clone().unwrap();
-        assert!(registry.is_running(&first_terminal_id));
-        assert!(registry.is_running(&second_terminal_id));
-
-        registry.kill(&first_terminal_id);
-        registry.kill(&second_terminal_id);
-    }
-
-    // ── shared GET helper (slice 3b-2) ──────────────────────────────────
-
-    async fn get(router: Router, uri: &str, auth: bool) -> (StatusCode, Value) {
-        let mut req = Request::builder().method("GET").uri(uri);
-        if auth {
-            req = req.header("x-auth-token", "tok");
-        }
-        let resp = router
-            .oneshot(req.body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-        let status = resp.status();
-        (status, body_json(resp).await)
-    }
-
-    // ── tabs/has ─────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn tabs_has_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = get(app(state), "/api/tabs/has?target=nope", false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn tabs_has_false_for_missing_target() {
-        let state = state_with_registry();
-        let (status, body) = get(app(state), "/api/tabs/has", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["exists"], json!(false));
-    }
-
-    #[tokio::test]
-    async fn tabs_has_true_for_known_tab_id() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, _pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = get(router, &format!("/api/tabs/has?target={tab_id}"), true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["exists"], json!(true));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    #[tokio::test]
-    async fn tabs_has_false_for_unknown_tab_id() {
-        let state = state_with_registry();
-        let (status, body) = get(app(state), "/api/tabs/has?target=does-not-exist", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["exists"], json!(false));
-    }
-
-    // ── tabs next/prev (honest deferral) ────────────────────────────────
-
-    #[tokio::test]
-    async fn tabs_next_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/tabs/next", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn tabs_next_is_honest_400_deferral() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/tabs/next", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        let msg = body["message"].as_str().unwrap();
-        assert!(msg.contains("ordered tab sequence"), "{msg}");
-    }
-
-    #[tokio::test]
-    async fn tabs_prev_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/tabs/prev", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn tabs_prev_is_honest_400_deferral() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/tabs/prev", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        let msg = body["message"].as_str().unwrap();
-        assert!(msg.contains("ordered tab sequence"), "{msg}");
-    }
-
-    // ── layout/snapshot ──────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn layout_snapshot_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = get(app(state), "/api/layout/snapshot", false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn layout_snapshot_empty_state_has_legacy_exact_top_level_keys() {
-        let state = state_with_registry();
-        let (status, body) = get(app(state), "/api/layout/snapshot", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let data = &body["data"];
-        assert_eq!(data["tabs"], json!([]));
-        assert!(data["activeTabId"].is_null());
-        assert_eq!(data["layouts"], json!({}));
-        assert_eq!(data["activePane"], json!({}));
-        assert_eq!(data["paneTitles"], json!({}));
-        assert_eq!(data["paneTitleSetByUser"], json!({}));
-    }
-
-    #[tokio::test]
-    async fn layout_snapshot_single_pane_tab_is_a_real_leaf_node() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = get(router, "/api/layout/snapshot", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let data = &body["data"];
-        assert_eq!(data["tabs"][0]["id"], json!(tab_id));
-        let leaf = &data["layouts"][&tab_id];
-        assert_eq!(leaf["type"], json!("leaf"));
-        assert_eq!(leaf["id"], json!(pane_id));
-        assert_eq!(leaf["content"]["kind"], json!("terminal"));
-        assert_eq!(leaf["content"]["terminalId"], json!(terminal_id));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    #[tokio::test]
-    async fn layout_snapshot_multi_pane_tab_is_an_honest_unknown_marker_not_a_fabricated_split() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_id, first_pane_id, first_terminal_id) = create_shell_tab(router.clone()).await;
-
-        let tmp = std::env::temp_dir();
-        let (status, split_body) = post(
-            router.clone(),
-            &format!("/api/panes/{first_pane_id}/split"),
-            json!({ "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{split_body}");
-        let second_pane_id = split_body["data"]["paneId"].as_str().unwrap().to_string();
-        let second_terminal_id = split_body["data"]["terminalId"]
-            .as_str()
-            .unwrap()
-            .to_string();
-
-        let (status, body) = get(router, "/api/layout/snapshot", true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let node = &body["data"]["layouts"][&tab_id];
-        assert_eq!(node["type"], json!("unknown"));
-        let mut ids: Vec<String> = node["paneIds"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|v| v.as_str().unwrap().to_string())
-            .collect();
-        ids.sort();
-        let mut expected = vec![first_pane_id.clone(), second_pane_id.clone()];
-        expected.sort();
-        assert_eq!(ids, expected);
-
-        let registry = state.terminal_registry.clone().unwrap();
-        registry.kill(&first_terminal_id);
-        registry.kill(&second_terminal_id);
-    }
-
-    #[tokio::test]
-    async fn layout_snapshot_tab_id_filter_narrows_to_one_tab() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (tab_a, _pane_a, terminal_a) = create_shell_tab(router.clone()).await;
-        let (_tab_b, _pane_b, terminal_b) = create_shell_tab(router.clone()).await;
-
-        let (status, body) =
-            get(router, &format!("/api/layout/snapshot?tabId={tab_a}"), true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let tabs = body["data"]["tabs"].as_array().unwrap();
-        assert_eq!(tabs.len(), 1, "{body}");
-        assert_eq!(tabs[0]["id"], json!(tab_a));
-
-        let registry = state.terminal_registry.clone().unwrap();
-        registry.kill(&terminal_a);
-        registry.kill(&terminal_b);
-    }
-
-    // ── navigate ─────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn navigate_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/navigate", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn navigate_pane_missing_url_is_400() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/panes/nope/navigate", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        assert_eq!(body["message"], json!("url required"));
-    }
-
-    #[tokio::test]
-    async fn navigate_unknown_pane_is_404() {
-        let state = state_with_registry();
-        let (status, body) = post(
-            app(state),
-            "/api/panes/does-not-exist/navigate",
-            json!({ "url": "https://example.com" }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::NOT_FOUND, "{body}");
-        assert_eq!(body["message"], json!("pane not found"));
-    }
-
-    #[tokio::test]
-    async fn navigate_pane_success_sets_browser_content_and_broadcasts_pane_attach() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-        let _ = rx.recv().await; // drain tab.create
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/navigate"),
-            json!({ "url": "https://example.com" }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["message"], json!("navigate requested"));
-
-        let frame = rx.recv().await.expect("pane.attach broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("pane.attach"));
-        assert_eq!(msg["payload"]["tabId"], json!(tab_id));
-        assert_eq!(msg["payload"]["paneId"], json!(pane_id));
-        assert_eq!(msg["payload"]["content"]["kind"], json!("browser"));
-        assert_eq!(
-            msg["payload"]["content"]["url"],
-            json!("https://example.com")
-        );
-
-        assert!(state.content_panes.lock().unwrap().get(&pane_id).is_some());
-        assert!(!state.terminal_panes.lock().unwrap().contains_key(&pane_id));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    // ── respawn ──────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn respawn_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/respawn", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn respawn_unknown_pane_is_404() {
-        let state = state_with_registry();
-        let (status, body) = post(
-            app(state),
-            "/api/panes/does-not-exist/respawn",
-            json!({}),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::NOT_FOUND, "{body}");
-        assert_eq!(body["message"], json!("pane not found"));
-    }
-
-    #[tokio::test]
-    async fn respawn_pane_replaces_terminal_in_place_and_broadcasts_pane_attach() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, pane_id, old_terminal_id) = create_shell_tab(router.clone()).await;
-        let _ = rx.recv().await; // drain tab.create
-
-        let tmp = std::env::temp_dir();
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/respawn"),
-            json!({ "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        let new_terminal_id = body["data"]["terminalId"].as_str().unwrap().to_string();
-        assert_ne!(new_terminal_id, old_terminal_id);
-
-        let frame = rx.recv().await.expect("pane.attach broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("pane.attach"));
-        assert_eq!(msg["payload"]["tabId"], json!(tab_id));
-        assert_eq!(msg["payload"]["paneId"], json!(pane_id));
-        assert_eq!(
-            msg["payload"]["content"]["terminalId"],
-            json!(new_terminal_id)
-        );
-
-        // Bookkeeping now points the SAME pane id at the NEW terminal --
-        // "replace in place", not a second pane.
-        assert_eq!(
-            state
-                .terminal_panes
-                .lock()
-                .unwrap()
-                .get(&pane_id)
-                .unwrap()
-                .terminal_id,
-            new_terminal_id
-        );
-
-        // Old terminal is orphaned-from-this-pane but still running in the
-        // shared registry (detach, don't kill -- this module's documented
-        // PTY-cleanup-parity finding, which this route also honors).
-        let registry = state.terminal_registry.clone().unwrap();
-        assert!(registry.is_running(&old_terminal_id));
-        assert!(registry.is_running(&new_terminal_id));
-
-        registry.kill(&old_terminal_id);
-        registry.kill(&new_terminal_id);
-    }
-
-    // ── attach (honest deferral) ─────────────────────────────────────────
-
-    #[tokio::test]
-    async fn attach_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/attach", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn attach_pane_is_honest_400_deferral() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/panes/nope/attach", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        let msg = body["message"].as_str().unwrap();
-        assert!(msg.contains("TerminalIdentityRegistry"), "{msg}");
-    }
-
-    // ── resize (honest deferral) ─────────────────────────────────────────
-
-    #[tokio::test]
-    async fn resize_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/resize", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn resize_pane_is_honest_400_deferral() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/panes/nope/resize", json!({}), true).await;
-        assert_eq!(status, StatusCode::BAD_REQUEST, "{body}");
-        let msg = body["message"].as_str().unwrap();
-        assert!(msg.contains("splitId"), "{msg}");
-        assert!(msg.contains("ui.layout.sync"), "{msg}");
-    }
-
-    // ── swap ─────────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn swap_pane_requires_auth() {
-        let state = state_with_registry();
-        let (status, _) = post(app(state), "/api/panes/nope/swap", json!({}), false).await;
-        assert_eq!(status, StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn swap_pane_missing_target_is_approx() {
-        let state = state_with_registry();
-        let (status, body) = post(app(state), "/api/panes/nope/swap", json!({}), true).await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["status"], json!("approx"));
-        assert_eq!(body["message"], json!("swap target missing"));
-    }
-
-    #[tokio::test]
-    async fn swap_unknown_pane_is_404() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            "/api/panes/does-not-exist/swap",
-            json!({ "target": pane_id }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::NOT_FOUND, "{body}");
-        assert_eq!(body["message"], json!("pane not found"));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    #[tokio::test]
-    async fn swap_unknown_other_is_404() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_id, pane_id, terminal_id) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_id}/swap"),
-            json!({ "target": "does-not-exist" }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::NOT_FOUND, "{body}");
-        assert_eq!(body["message"], json!("pane not found"));
-
-        state.terminal_registry.clone().unwrap().kill(&terminal_id);
-    }
-
-    #[tokio::test]
-    async fn swap_cross_tab_panes_reports_panes_not_found() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let (_tab_a, pane_a, terminal_a) = create_shell_tab(router.clone()).await;
-        let (_tab_b, pane_b, terminal_b) = create_shell_tab(router.clone()).await;
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{pane_a}/swap"),
-            json!({ "target": pane_b }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["message"], json!("panes not found"));
-
-        let registry = state.terminal_registry.clone().unwrap();
-        registry.kill(&terminal_a);
-        registry.kill(&terminal_b);
-    }
-
-    #[tokio::test]
-    async fn swap_two_terminal_panes_in_same_tab_exchanges_bookkeeping_and_broadcasts() {
-        let state = state_with_registry();
-        let router = app(state.clone());
-        let mut rx = state.broadcast_tx.subscribe();
-        let (tab_id, first_pane_id, first_terminal_id) = create_shell_tab(router.clone()).await;
-        let _ = rx.recv().await; // drain tab.create
-
-        let tmp = std::env::temp_dir();
-        let (status, split_body) = post(
-            router.clone(),
-            &format!("/api/panes/{first_pane_id}/split"),
-            json!({ "cwd": tmp.to_string_lossy() }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{split_body}");
-        let second_pane_id = split_body["data"]["paneId"].as_str().unwrap().to_string();
-        let second_terminal_id = split_body["data"]["terminalId"]
-            .as_str()
-            .unwrap()
-            .to_string();
-        let _ = rx.recv().await; // drain pane.split
-
-        let (status, body) = post(
-            router,
-            &format!("/api/panes/{first_pane_id}/swap"),
-            json!({ "target": second_pane_id }),
-            true,
-        )
-        .await;
-        assert_eq!(status, StatusCode::OK, "{body}");
-        assert_eq!(body["data"]["tabId"], json!(tab_id));
-        assert_eq!(body["message"], json!("panes swapped"));
-
-        let frame = rx.recv().await.expect("pane.swap broadcast");
-        let msg: Value = serde_json::from_str(&frame).unwrap();
-        assert_eq!(msg["command"], json!("pane.swap"));
-        assert_eq!(msg["payload"]["tabId"], json!(tab_id));
-        assert_eq!(msg["payload"]["paneId"], json!(first_pane_id));
-        assert_eq!(msg["payload"]["otherId"], json!(second_pane_id));
-
-        // Bookkeeping exchanged: first pane id now owns the SECOND terminal
-        // and vice versa.
-        let terminal_panes = state.terminal_panes.lock().unwrap();
-        assert_eq!(
-            terminal_panes.get(&first_pane_id).unwrap().terminal_id,
-            second_terminal_id
-        );
-        assert_eq!(
-            terminal_panes.get(&second_pane_id).unwrap().terminal_id,
-            first_terminal_id
-        );
-        drop(terminal_panes);
-
-        let registry = state.terminal_registry.clone().unwrap();
-        registry.kill(&first_terminal_id);
-        registry.kill(&second_terminal_id);
-    }
-}
+#[path = "pane_ops_tests.rs"]
+mod tests;
+
+#[cfg(test)]
+#[path = "pane_ops_tab_tests.rs"]
+mod tab_tests;
